refactor: (#31) components index.ts 참조하도록 변경
