+++ src/components/Advertisement.tsx
@@ -3,7 +3,7 @@ interface AdvertisementProps {
   smallAdSrc: string;
 }
 
-export const Advertisement = ({ largeAdSrc, smallAdSrc }: AdvertisementProps) => {
+const Advertisement = ({ largeAdSrc, smallAdSrc }: AdvertisementProps) => {
   return (
     <div className="mt-20 w-1/5 space-y-4">
       {largeAdSrc && <img src={largeAdSrc} alt="대형 광고" />}
@@ -11,3 +11,5 @@ export const Advertisement = ({ largeAdSrc, smallAdSrc }: AdvertisementProps) =>
     </div>
   );
 };
+
+export default Advertisement;
